Refactor theme management by removing deprecated theme-constants file
- Deleted the `theme-constants.ts` file to streamline theme imports.
- Updated references in `_layout.tsx` and `android-navigation-bar.ts` to use the new theme structure.
- This change enhances code organization and maintains consistency in theme usage across the application.
